Added scripts via upload
We had RNA-seq data from another publication that we wanted to compare against the germ cell QTL data. Upon knockout of the QTL region, the group had analyzed transcriptome differences. Since we had QTL data in our lab, we wanted to look that whether the DE genes overlapped with our QTL targets or not. We made use of the RNA-seq data and later analyzed whether the DE genes overlapped with QTL targets or not.